+++ node/src/main/kotlin/net/corda/node/internal/CordaRPCOpsImpl.kt
@@ -12,7 +12,6 @@ import net.corda.core.identity.Party
 import net.corda.core.messaging.*
 import net.corda.core.node.NodeInfo
 import net.corda.core.node.services.NetworkMapCache
-import net.corda.core.node.services.StateMachineTransactionMapping
 import net.corda.core.node.services.Vault
 import net.corda.core.node.services.vault.PageSpecification
 import net.corda.core.node.services.vault.QueryCriteria
@@ -76,7 +75,7 @@ class CordaRPCOpsImpl(
 
     override fun verifiedTransactionsFeed(): DataFeed<List<SignedTransaction>, SignedTransaction> {
         return database.transaction {
-            services.storageService.validatedTransactions.track()
+            services.validatedTransactions.track()
         }
     }
 
@@ -92,7 +91,7 @@ class CordaRPCOpsImpl(
 
     override fun stateMachineRecordedTransactionMappingFeed(): DataFeed<List<StateMachineTransactionMapping>, StateMachineTransactionMapping> {
         return database.transaction {
-            services.storageService.stateMachineRecordedTransactionMapping.track()
+            services.stateMachineRecordedTransactionMapping.track()
         }
     }
 
@@ -143,21 +142,21 @@ class CordaRPCOpsImpl(
     override fun attachmentExists(id: SecureHash): Boolean {
         // TODO: this operation should not require an explicit transaction
         return database.transaction {
-            services.storageService.attachments.openAttachment(id) != null
+            services.attachments.openAttachment(id) != null
         }
     }
 
     override fun openAttachment(id: SecureHash): InputStream {
         // TODO: this operation should not require an explicit transaction
         return database.transaction {
-            services.storageService.attachments.openAttachment(id)!!.open()
+            services.attachments.openAttachment(id)!!.open()
         }
     }
 
     override fun uploadAttachment(jar: InputStream): SecureHash {
         // TODO: this operation should not require an explicit transaction
         return database.transaction {
-            services.storageService.attachments.importAttachment(jar)
+            services.attachments.importAttachment(jar)
         }
     }
 
@@ -166,7 +165,7 @@ class CordaRPCOpsImpl(
     override fun currentNodeTime(): Instant = Instant.now(services.clock)
     @Suppress("OverridingDeprecatedMember", "DEPRECATION")
     override fun uploadFile(dataType: String, name: String?, file: InputStream): String {
-        val acceptor = services.storageService.uploaders.firstOrNull { it.accepts(dataType) }
+        val acceptor = services.uploaders.firstOrNull { it.accepts(dataType) }
         return database.transaction {
             acceptor?.upload(file) ?: throw RuntimeException("Cannot find file upload acceptor for $dataType")
         }